Removed react-dom (This was not a web application) Added react-redux package. Created root reducer/user reducer. Restructured Navigation away, and moved initialized store in App.js
