+++ package.json
@@ -25,12 +25,12 @@
     "moment": "^2.22.2",
     "path": "^0.12.7",
     "react": "^16.3.1",
-    "react-dom": "^16.4.2",
     "react-native": "https://github.com/expo/react-native/archive/sdk-27.0.0.tar.gz",
     "react-native-elements": "^0.19.1",
     "react-native-gifted-chat": "^0.4.3",
     "react-native-platform-touchable": "^1.1.1",
     "react-navigation": "^2.3.1",
+    "react-redux": "^5.0.7",
     "redux": "^4.0.0",
     "util": "*"
   }
