+++ rootReducer.js
@@ -1,0 +1,6 @@
+import userReducer from './src/reducers/user';
+import { combineReducers } from 'redux';
+
+export default combineReducers({
+    user: userReducer
+});
